+++ src/main/java/eu/dissco/core/handlemanager/component/PidResolver.java
@@ -1,4 +1,4 @@
-package eu.dissco.core.handlemanager.web;
+package eu.dissco.core.handlemanager.component;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
