+++ src/test/java/eu/dissco/core/handlemanager/service/FdoRecordServiceTest.java
@@ -62,6 +62,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 
 import eu.dissco.core.handlemanager.Profiles;
+import eu.dissco.core.handlemanager.component.PidResolver;
 import eu.dissco.core.handlemanager.domain.fdo.FdoProfile;
 import eu.dissco.core.handlemanager.domain.fdo.FdoType;
 import eu.dissco.core.handlemanager.domain.repsitoryobjects.FdoAttribute;
@@ -72,7 +73,6 @@ import eu.dissco.core.handlemanager.exceptions.PidResolutionException;
 import eu.dissco.core.handlemanager.properties.ApplicationProperties;
 import eu.dissco.core.handlemanager.schema.HandleRequestAttributes;
 import eu.dissco.core.handlemanager.schema.OtherspecimenIds;
-import eu.dissco.core.handlemanager.web.PidResolver;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
